modems bug 99: try to ignore usb0 exposed by AC341U
The AC341U ends up exposing a usb0 (rndis_host) and a wwan0
(qmi_wwan). In order for the modem management scripts to work
properly, we need to completely ignore the usb0.

We do this by tagging 1199:9057 as qmiwwan in the hybrid modems list,
and then checking the type against the "qmi_wwan" driver. If that
doesn't match (as it wouldn't with the usb0 interface), we ignore
it. Note that this logic won't affect the WWAN interface exposed by
the modem as that is actually finished detecting support just by
matching driver against "qmi_wwan" in the previous logic block.

(cherry picked from commit ae06a8ca573aae60ddb74a01190b532d24dee467)