re #18281 Boot command and package fixes for Hyper-V/Azure
Xen and Hyper-V don't need the console to be specified as "hvc0".
In fact, ttyS0 works just great, on Hyper-V, Xen and AWS.

Also, made the hyper-v modules not auto-load (which fails on all the
other (non-hyperV) platforms).  Instead, vcearlyinit will load the
hyper-V drivers after verifying that the VM is running on hyper-V.
(Also applies to Azure).