re #18676 pciutils: do not update pci.ids during the build
Checked in a recent pci.ids into .../pciutils/files, and use that to
update the copy in the pciutils tarball before the build.  Added a
README.txt explaining how to update the pci.ids file periodically.

Checking in the uncompressed pci.ids file, so that we can maintain a
sensible log of what is changing.

@review egorovv