re #17689 IGB edge5x0: wait for MDIO to become ready at system startup
If we set loglevel=6 on the kernel cmdline, the system boots up fast
enough that the MDIO bus doesn't seem to fully initialize, and the
initial probe from the IGB driver for our switch sometimes fails.

Sandra's fix: loop around for a bit until MDIO is ready, and then continue.

review @sberndt