re #19166 Handle old-style network-interfaces config from meta-data
We have already published to some customers the old (pre-0.7.9) way of
passing in network interface configuration via meta-data (the network-interfaces
key, with a custom dict that is openwrt-specific).

In order to support both the new style (the network-config top-level file) as
well as the old one, we move the conversion of the "eni"-style data (the old
style "network-interfaces" blob) into the distro class with a default, and
override it in velocloud.py.

PS The velocloud code now supports both old styles (the old dictionary style,
as well as the old Ubuntu /etc/network/interfaces-style blob).

review @vcalex