modems bug 123: support GH-UDG-MCLTEC modem
This is a modem for the Japanese market that provides a built-in web
interface for management. It works in router mode, with 192.168.225.x
as local IP address in the network interface once DHCP has been run on
it.

(cherry picked from commit 471323858836a0cacad1ea8a9675f97a109b5f7f)