modems: support USB730L
Explicitly support the USB730L as an auto-connected device.

There are multiple versions of this device with different USB
layouts. One of the versions comes with a standard ECM interface that
is just used as an autoconnected modem. A different version comes with
a non-standard RNDIS interface that needs a custom kernel patch (not
yet implemented).

This patch adds support for at least the ECM based device.

(cherry picked from commit 0fcce944694ab51b8a3ae1306a9096ef9777f87a)