Fix imagebuilder breakage caused by previous checkin for 510 firmware blobs
The symlinks under target/linux/x64/edge5x0/files point to files outside
the "trunk" (base of the openwrt tree).  The Makefile attempts to copy
from that location to the target directory, but that command gets executed
both in the original openwrt build, as well as whenever the imagebuilder
itself is run somewhere else (where those symlinks aren't valid).

Fixed by (a) making the cp failure ignore errors in the image Makefile,
and (b) ensuring that the imagebuilder contains the real firmware blobs
in the target directory, so that the image creation script can find them.

review @sberndt