modems bug 125: don't leave subshells around when stopping
We were processing each modem path in a separate subshell, as the
while() loop was processing lines one by one. This was preventing a
proper shutdown because killing the parent process would still leave
the subshell running.

So just avoid creating the subshells; get the whole list of DBus paths
to check and process them one by one after that.