modemmanager: bump git snapshot
This new version includes a bunch of stability fixes when
udev isn't being used (e.g. when the kernel events are
reported via mmcli).

In particular, the new version avoids invalid memory R/W
as well as segfaults when a QMI modem gets removed in the
middle of the device probing or initialization steps (as
happening with the SkyUS DS modem in Case 170808-11717).