ENG-20412 Power control in KVM and Supermicro edges
KVM virtual machines, and the Supermicro edges (840, 1000, 2000)
generate ACPI Power events when the power button is pressed once
(or the KVM equivalent operation of "Shutdown VM" - as opposed to
"Hard Power off", which is like a long-press of the hardware power
button).

Added an "acpid" busybox daemon to the edge, to respond to these
ACPI power events, and power down the edge gracefully (instead of
having to long-press the button, or invoke Hard Power Off for the
VM, and shut it down abruptly, which can corrupt filesystems).