modems bug 126: don't flag 'wwanX' named interfaces as qmiwwan type
The logic was matching the 'wwanX' name to a "qmiwwan" type modem,
which is wrong. The 'wwanX' name may be assigned by the kernel to
network interfaces exposed by multiple drivers, including:
  * qmi_wwan
  * cdc-mbim
  * cdc-ncm with FLAG_WWAN
  * cdc-ether with FLAG_WWAN

In our case, we prefer to assign it the "cdcether" type instead, as
the QMI modems would already be matches automatically by driver just
before. Once matched as "cdcether", the hybridmodemslist allows
overriding this selection (e.g. for qmihybrid modems).