ENG-20879 Integrate FRR 3.0 branch into openwrt toolchain
Integrating FRR's master is a major task (because of badly broken
cross-compilation rules for a new python plugin).  In the meantime,
I'm porting the "3.0" branch of FRR here, to start with, and enabling
only PIMD within that.

(This required updates to a few package dependencies, notably popt,
automake and autoconf.  This FRR makefile, pointed to from the docs
directory of the FRR master branch, only works with automake 1.15.1
and later.)

After this, we can create a branch based off the 3.0 FRR branch for
VeloCloud changes, that we can pull in here by bumping up the commit
ID in trunk/package/frr/Makefile.

Once the dust settles a bit, we will try to tackle the master branch
(either by somehow brute-forcing the cross-compilation setup to
work, or writing a pure-python substitute for the native module, or
something..)