Update gdb to 7.9.1
The old one havd a problem with vfork-based subprocess invocation
rendering attached debugging unusable.
This version has a different problem fixed in the later releases but 8.x
needs a C++11 capable compiler to build so this is the end of the road.
With the supplied patch borrowed from upstream it seem to work for me.