More image generation changes: switch installer generation to make_ext4fs
We also need to generate the installer itself with make_ext4fs, to
allow for jumbo installer images (e.g. with 6+ GB of VNF images).
Allow additional directories to be passed in to the installer
creation via the environment variable ENV_ADDITIONAL_INST_DIRS;
these will end up under the "images" directory in the "inst" partition.

Also fixed the imagebuilder creation, to include the mfgtest binaries
properly (so that they can be copied to the final installers).